Full compilation with generics and specailisation of Photon SIDM4

+++ swift/mc/mc/BulkHenyeyGreenstein.swift
@@ -29,9 +29,9 @@ class BulkHenyeyGreenstein<T, V, M:MatrixProtocol> : BulkMaterial<T, V, M> where
                 let den = 1.0 - g + 2.0 * randomfloat() * g
                 let rand_frac =  num / den
                 let frac = (1.0 + g*g - rand_frac*rand_frac) / T(2.0*g)
-                θ = T(acos( Double(frac) ))
+                θ = T(acos(Double(frac) ))
             } else {
-                θ = T(acos(Float(1.0 - 2.0 * randomfloat() )))
+                θ = T(acos(Double(1.0 - 2.0 * randomfloat() )))
             }
             if θ >= 0 && θ <= T.pi {
                 break
